move more files to shared
