+++ frontend/src/modules/controllers/index.ts
@@ -11,13 +11,11 @@
  *
  * Re-usable controller components.
  */
+export * from './ValueDisplay';
+export * from './ValueSelectorDisplay';
 export { default as ValueSlider } from './ValueSlider';
-export { default as TabPanel } from './TabPanel';
-export { default as TestTool } from './TestTool';
 export { default as ToogleValue } from './ToggleValue';
 export { default as ValueModal } from './ValueModal';
 export { default as ValueController } from './ValueClicker';
 export { default as ValueClicker } from './ValueClicker';
-export { default as SimpleTable } from './SimpleTable';
-export * from './ValueDisplay';
-export * from './ValueSelectorDisplay';
+export { default as SimpleTable } from '../shared/SimpleTable';
