+++ frontend/src/modules/shared/SimpleTable.tsx
