+++ frontend/src/modules/shared/index.ts
@@ -1,2 +1,4 @@
-export { default as Knob } from './Knob';
 export * from './ModalPopup';
+export { default as Knob } from './Knob';
+export { default as TabPanel } from './TabPanel';
+export { default as TestTool } from './TestTool';
